refactor: move cad-specific stuff out of core
