+++ viewer/core/package.json
@@ -10,12 +10,14 @@
   },
   "dependencies": {
     "@reveal/cad-geometry-loaders": "workspace:*",
+    "@reveal/cad-model": "workspace:*",
     "@reveal/cad-parsers": "workspace:*",
     "@reveal/cad-styling": "workspace:*",
     "@reveal/camera-manager": "workspace:*",
     "@reveal/data-source": "workspace:*",
     "@reveal/logger": "workspace:*",
     "@reveal/metrics": "workspace:*",
+    "@reveal/model-base": "workspace:*",
     "@reveal/modeldata-api": "workspace:*",
     "@reveal/nodes-api": "workspace:*",
     "@reveal/rendering": "workspace:*",
