--- viewer/core/src/datamodels/base/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright 2021 Cognite AS
- */
-
-export { MetadataRepository } from '@reveal/cad-parsers';
-export * from './SupportedModelTypes';
-export * from './types';
