+++ viewer/core/src/public/migration/Cognite3DViewer.ts
@@ -15,7 +15,7 @@ import { assertNever, EventTrigger, InputHandler, disposeOfAllEventListeners } f
 import { MetricsLogger } from '@reveal/metrics';
 
 import { worldToNormalizedViewportCoordinates, worldToViewportCoordinates } from '../../utilities/worldToViewport';
-import { intersectCadNodes } from '../../datamodels/cad/picking';
+import { intersectCadNodes } from '../../../../packages/cad-model/src/picking';
 
 import {
   AddModelOptions,
@@ -35,19 +35,18 @@ import { RevealManager } from '../RevealManager';
 import { DisposedDelegate, SceneRenderedDelegate } from '../types';
 
 import { Spinner } from '../../utilities/Spinner';
-
-import { IntersectInput, SupportedModelTypes } from '../../datamodels/base';
 import { intersectPointClouds } from '../../datamodels/pointcloud/picking';
 
 import { CadIntersection, IntersectionFromPixelOptions, PointCloudIntersection, RevealOptions } from '../..';
 import { PropType } from '../../utilities/reflection';
-import { CadModelSectorLoadStatistics } from '../../datamodels/cad/CadModelSectorLoadStatistics';
+import { CadModelSectorLoadStatistics } from '../../../../packages/cad-model/src/CadModelSectorLoadStatistics';
 import { ViewerState, ViewStateHelper } from '../../utilities/ViewStateHelper';
 import { RevealManagerHelper } from '../../storage/RevealManagerHelper';
 
 import { DefaultCameraManager, CameraManager } from '@reveal/camera-manager';
 import { CdfModelIdentifier, File3dFormat } from '@reveal/modeldata-api';
 import { DataSource, CdfDataSource, LocalDataSource } from '@reveal/data-source';
+import { IntersectInput, SupportedModelTypes } from '@reveal/model-base';
 
 import { CogniteClient } from '@cognite/sdk';
 import log from '@reveal/logger';
