+++ viewer/core/src/public/migration/CogniteModelBase.ts
@@ -3,7 +3,7 @@
  */
 
 import { CameraConfiguration } from './types';
-import { SupportedModelTypes } from '../../datamodels/base';
+import { SupportedModelTypes } from '@reveal/model-base';
 
 /**
  * Base class for 3D models supported by {@link Cognite3DViewer}.
