+++ viewer/packages/cad-geometry-loaders/app/index.ts
@@ -4,12 +4,12 @@
 
 import * as THREE from 'three';
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls';
-import { CadModelFactory } from '../../../core/src/datamodels/cad/CadModelFactory';
+import { CadModelFactory } from '../../cad-model/src/CadModelFactory';
 import { CadMaterialManager } from '@reveal/rendering';
 import { CdfModelDataProvider, CdfModelIdentifier, CdfModelMetadataProvider } from '@reveal/modeldata-api';
 import { CadModelUpdateHandler } from '../src/CadModelUpdateHandler';
 import { ByScreenSizeSectorCuller } from '..';
-import { CadManager } from '../../../core/src/datamodels/cad/CadManager';
+import { CadManager } from '../../cad-model/src/CadManager';
 import { revealEnv } from '@reveal/utilities';
 import dat from 'dat.gui';
 import { createApplicationSDK } from '../../../test-utilities/src/appUtils';
