+++ viewer/packages/cad-model/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright 2022 Cognite AS
+ */
+
+export { CadManager } from './src/CadManager';
+export { GeometryFilter } from './src/GeometryFilter';
+export { CadModelSectorLoadStatistics } from './src/CadModelSectorLoadStatistics';
