+++ viewer/packages/cad-model/src/CadModelSectorLoadStatistics.ts
