+++ viewer/packages/cad-model/src/picking.test.ts
@@ -7,8 +7,8 @@ import * as THREE from 'three';
 import { CadNode } from '@reveal/rendering';
 
 import { intersectCadNodes } from './picking';
-import { IntersectInput } from '../base/types';
-import { createGlContext } from '../../../../test-utilities';
+import { IntersectInput } from '../../model-base/src/types';
+import { createGlContext } from '../../../test-utilities';
 
 describe('intersectCadNodes', () => {
   const camera = new THREE.PerspectiveCamera();
