+++ viewer/packages/cad-parsers/src/metadata/CadModelClipper.ts
