+++ viewer/packages/model-base/index.ts
@@ -1,0 +1,6 @@
+/*!
+ * Copyright 2022 Cognite AS
+ */
+
+export { IntersectInput } from './src/types';
+export { SupportedModelTypes } from './src/SupportedModelTypes';
