+++ viewer/packages/model-base/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@reveal/model-base",
+  "private": true,
+  "main": "index.ts",
+  "scripts": {
+    "start": "yarn ws:start",
+    "test": "yarn ws:test --config ./../../jest.config.js",
+    "lint": "yarn ws:lint"
+  }
+}
