+++ viewer/packages/model-base/src/SupportedModelTypes.ts
